A fix: - contact header transport was always UDP ... in sipcommonconfig, the transport was not correctly set (as it now can be   for each identity, it had to be taken from the proxy->transport ... and it wasn't).
A compilation warning fix:
- in IM message ... just a warning.

Still, the whole TLS stuff is not really good. 
A  new socket is created per each message sent ... 
and when the phone is behind a NAT, i don't seem to be able to get incoming messages ... but this may have to do with 
my ser config file ... i'll get back to it ...



git-svn-id: svn://svn.minisip.org/minisip/trunk@1687 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/messages/SipIMMessage.cxx
@@ -69,7 +69,7 @@ SipIMMessage::SipIMMessage(string branch,
 	this->toUri = toUri;
 	int posAt = toUri.find("@");
 	toUser = toUri.substr( 0, posAt );
-	if( posAt != string::npos ) 
+	if( posAt != (int)string::npos ) 
 		toDomain = toUri.substr( posAt + 1 );
 	else 	toDomain = "";
